refactor(settings_ui): lift backup snapshot + config handlers to _handlers_backups.py
Snapshot ops (list/view/diff/restore/delete/bulk) + auto-backup config
(get/save). Decomposes _validate_backup_payload (C901 20 -> under) via
per-type coercers + _BACKUP_INT_BOUNDS, and _save_backup_config (11 ->
under) via _save_backup_config_addon / _filter_env_pinned_backup_fields.
