+++ src/ha_mcp/settings_ui/__init__.py
@@ -38,6 +38,7 @@ from ..errors import ErrorCode, create_error_response
 from ..transforms import DEFAULT_PINNED_TOOLS
 from ..utils.data_paths import get_data_dir
 from . import _persistence, _supervisor
+from ._handlers_backups import build_backups_handlers
 from ._handlers_fs import build_fs_handlers
 from ._handlers_theme import build_theme_handlers
 from ._handlers_tools import build_tools_handlers
@@ -1194,189 +1195,6 @@ def build_settings_handlers(
 
     # ---- Auto-backup routes (#1288) ----
 
-    def _backup_mgr() -> Any:
-        settings = get_global_settings()
-        client = getattr(server, "client", None) or getattr(server, "_client", None)
-        return get_backup_manager(client, settings) if client is not None else None
-
-    def _bad_request(
-        message: str,
-        *,
-        code: ErrorCode = ErrorCode.VALIDATION_INVALID_PARAMETER,
-        status: int = 400,
-    ) -> JSONResponse:
-        return JSONResponse(create_error_response(code, message), status_code=status)
-
-    def _not_found(name: str) -> JSONResponse:
-        return JSONResponse(
-            create_error_response(
-                ErrorCode.RESOURCE_NOT_FOUND, f"Backup {name!r} not found"
-            ),
-            status_code=404,
-        )
-
-    async def _list_backups(request: Request) -> JSONResponse:
-        mgr = _backup_mgr()
-        if mgr is None:
-            return _bad_request("Backup manager unavailable")
-        params = request.query_params
-        try:
-            limit = int(params.get("limit", "500"))
-        except ValueError:
-            return _bad_request("'limit' must be an integer")
-        # Offload sync directory I/O to keep the request handler async-clean.
-        entries = await asyncio.to_thread(
-            mgr.list_snapshots,
-            domain=params.get("domain") or None,
-            entity_id=params.get("entity_id") or None,
-            limit=max(1, min(10_000, limit)),
-        )
-        settings = get_global_settings()
-        return JSONResponse(
-            {
-                "success": True,
-                "backups": entries,
-                "count": len(entries),
-                "backup_dir": str(mgr.backup_dir),
-                "enabled": mgr.enabled,
-                "throttle_minutes": settings.auto_backup_throttle_minutes,
-                "retain_per_entity": settings.auto_backup_retain_per_entity,
-            }
-        )
-
-    async def _view_backup(request: Request) -> JSONResponse:
-        mgr = _backup_mgr()
-        if mgr is None:
-            return _bad_request("Backup manager unavailable")
-        name = request.path_params.get("name", "")
-        try:
-            data = await asyncio.to_thread(mgr.read_snapshot, name)
-        except FileNotFoundError:
-            return _not_found(name)
-        except ValueError as err:
-            return _bad_request(str(err))
-        return JSONResponse({"success": True, "data": data})
-
-    async def _diff_backup(request: Request) -> JSONResponse:
-        import difflib
-
-        import yaml  # type: ignore[import-untyped]
-
-        mgr = _backup_mgr()
-        if mgr is None:
-            return _bad_request("Backup manager unavailable")
-        name = request.path_params.get("name", "")
-        try:
-            snapshot = await asyncio.to_thread(mgr.read_snapshot, name)
-        except FileNotFoundError:
-            return _not_found(name)
-        except ValueError as err:
-            return _bad_request(str(err))
-        handler = mgr.handler_for(snapshot["domain"])
-        if handler is None:
-            return _bad_request(
-                f"No handler for domain {snapshot['domain']!r}; cannot diff",
-                code=ErrorCode.RESOURCE_NOT_FOUND,
-                status=404,
-            )
-        client = getattr(server, "client", None) or getattr(server, "_client", None)
-        # Narrow to transport / HA-API / FS errors so programming bugs
-        # propagate to the request handler instead of decorating the diff
-        # output with a "_error" sentinel masquerading as entity state.
-        from ..backup_manager import _CAPTURE_TRANSIENT_ERRORS
-
-        try:
-            current = await handler.fetch(client, snapshot["entity_id"])
-        except _CAPTURE_TRANSIENT_ERRORS as err:
-            current = {"_error": f"{type(err).__name__}: {err}"}
-        backup_yaml = yaml.safe_dump(
-            snapshot.get("config"), default_flow_style=False, sort_keys=True
-        ).splitlines()
-        current_yaml = yaml.safe_dump(
-            current, default_flow_style=False, sort_keys=True
-        ).splitlines()
-        diff = list(
-            difflib.unified_diff(
-                backup_yaml,
-                current_yaml,
-                fromfile=f"backup:{name}",
-                tofile=f"current:{snapshot['entity_id']}",
-                lineterm="",
-            )
-        )
-        return JSONResponse(
-            {
-                "success": True,
-                "diff": "\n".join(diff),
-                "backup_present": current is not None
-                and not (isinstance(current, dict) and "_error" in current),
-            }
-        )
-
-    async def _restore_backup(request: Request) -> JSONResponse:
-        mgr = _backup_mgr()
-        if mgr is None:
-            return _bad_request("Backup manager unavailable")
-        name = request.path_params.get("name", "")
-        try:
-            result = await mgr.restore_snapshot(name)
-        except FileNotFoundError:
-            return _not_found(name)
-        except (ValueError, LookupError) as err:
-            return _bad_request(str(err))
-        return JSONResponse({"success": True, "data": result})
-
-    async def _delete_backup(request: Request) -> JSONResponse:
-        mgr = _backup_mgr()
-        if mgr is None:
-            return _bad_request("Backup manager unavailable")
-        name = request.path_params.get("name", "")
-        try:
-            await asyncio.to_thread(mgr.delete_snapshot, name)
-        except FileNotFoundError:
-            return _not_found(name)
-        except ValueError as err:
-            return _bad_request(str(err))
-        return JSONResponse({"success": True, "deleted": [name]})
-
-    async def _delete_backups_bulk(request: Request) -> JSONResponse:
-        mgr = _backup_mgr()
-        if mgr is None:
-            return _bad_request("Backup manager unavailable")
-        params = request.query_params
-        older = params.get("older_than_days")
-        try:
-            older_int = int(older) if older is not None else None
-        except ValueError:
-            return _bad_request("'older_than_days' must be an integer")
-        if (
-            params.get("domain") is None
-            and params.get("entity_id") is None
-            and older_int is None
-        ):
-            return _bad_request(
-                "Bulk delete requires at least one filter "
-                "(domain, entity_id, older_than_days)"
-            )
-        try:
-            bulk = await asyncio.to_thread(
-                mgr.delete_bulk,
-                domain=params.get("domain") or None,
-                entity_id=params.get("entity_id") or None,
-                older_than_days=older_int,
-            )
-        except ValueError as err:
-            return _bad_request(str(err))
-        return JSONResponse(
-            {
-                "success": True,
-                "deleted": bulk["deleted"],
-                "failed": bulk["failed"],
-                "count": len(bulk["deleted"]),
-                "failed_count": len(bulk["failed"]),
-            }
-        )
-
     async def _get_advanced_settings(_: Request) -> JSONResponse:
         """Return advanced (non-feature-flag, non-backup) settings + per-field
         origin + editable flag.
@@ -1877,232 +1695,6 @@ def build_settings_handlers(
             status_code=400,
         )
 
-    async def _get_backup_config(_: Request) -> JSONResponse:
-        """Return live auto-backup config + per-field origin + editable flag.
-
-        Per-field origin/editable matrix (see ``config.get_backup_setting_origin``):
-        - ``addon``: editable — POST routes through Supervisor.
-        - ``env``: read-only — env var wins; user must unset to edit.
-        - ``file``/``default``: editable — POST writes the override file.
-        """
-        settings = get_global_settings()
-        addon_mode = is_running_in_addon()
-        fields = []
-        for field_name, env_name, _ftype in BACKUP_OVERRIDE_FIELDS:
-            origin = get_backup_setting_origin(env_name)
-            editable = origin in ("addon", "file", "default")
-            fields.append(
-                {
-                    "field": field_name,
-                    "env_var": env_name,
-                    "value": getattr(settings, field_name),
-                    "origin": origin,
-                    "editable": editable,
-                }
-            )
-        return JSONResponse(
-            {
-                "success": True,
-                "is_addon": addon_mode,
-                "fields": fields,
-            }
-        )
-
-    def _validate_backup_payload(payload: Any) -> tuple[dict[str, Any], str | None]:
-        """Coerce and bounds-check the POST body. Returns (clean, error_msg)."""
-        if not isinstance(payload, dict):
-            return {}, "Body must be a JSON object"
-        clean: dict[str, Any] = {}
-        for field_name, _env_name, ftype in BACKUP_OVERRIDE_FIELDS:
-            if field_name not in payload:
-                continue
-            raw = payload[field_name]
-            if ftype is bool:
-                if isinstance(raw, bool):
-                    value: Any = raw
-                elif isinstance(raw, int):
-                    value = bool(raw)
-                elif isinstance(raw, str):
-                    s = raw.strip().lower()
-                    if s in ("true", "1", "yes", "on"):
-                        value = True
-                    elif s in ("false", "0", "no", "off"):
-                        value = False
-                    else:
-                        return {}, f"Invalid boolean for {field_name}: {raw!r}"
-                else:
-                    return {}, f"Invalid value for {field_name}: {raw!r}"
-            elif ftype is int:
-                if isinstance(raw, bool) or not isinstance(raw, int | str):
-                    return {}, f"Invalid integer for {field_name}: {raw!r}"
-                try:
-                    value = int(raw)
-                except (ValueError, TypeError):
-                    return {}, f"Invalid integer for {field_name}: {raw!r}"
-                if field_name == "auto_backup_throttle_minutes" and not (
-                    0 <= value <= 1440
-                ):
-                    return {}, "auto_backup_throttle_minutes must be 0..1440"
-                if field_name == "auto_backup_retain_per_entity" and not (
-                    1 <= value <= 10_000
-                ):
-                    return {}, "auto_backup_retain_per_entity must be 1..10000"
-                if field_name == "auto_backup_calendar_lookahead_days" and not (
-                    1 <= value <= 365
-                ):
-                    return (
-                        {},
-                        "auto_backup_calendar_lookahead_days must be 1..365",
-                    )
-            elif ftype is str:
-                if not isinstance(raw, str):
-                    return {}, f"Invalid string for {field_name}: {raw!r}"
-                if "\x00" in raw:
-                    return {}, f"{field_name} must not contain null bytes"
-                value = raw
-            else:
-                continue
-            clean[field_name] = value
-        if not clean:
-            return {}, "No editable auto-backup fields in body"
-        return clean, None
-
-    async def _save_backup_config(request: Request) -> JSONResponse:
-        """Persist auto-backup config edits and publish to the live process.
-
-        Routing:
-        - Addon mode: POST ``/addons/self/options`` (with the existing
-          options merged so required schema keys like ``backup_hint``
-          survive the full-replacement validation) and return
-          ``restart_required=True``. ``start.py`` will re-derive env
-          vars from ``config.yaml`` on the next addon boot, but the
-          actual restart is fired by the user clicking the global
-          Restart Add-on button — NOT by this handler. Same unified
-          flow as the Tools and Server Settings save endpoints.
-        - Standalone (file) mode: refuse any field that's pinned by an
-          env var (process or ``.env``) — return 409 with the offending
-          names so the UI can refresh and show the read-only banner.
-          Editable fields merge into
-          ``<data_dir>/backup_settings.json`` and a Settings cache
-          reset publishes them immediately, hence
-          ``restart_required=False``.
-        """
-        try:
-            payload = await request.json()
-        except (ValueError, json.JSONDecodeError):
-            return _bad_request("Invalid JSON body")
-        clean, err = _validate_backup_payload(payload)
-        if err is not None:
-            return _bad_request(err)
-
-        if is_running_in_addon():
-            # ``is_running_in_addon()`` checks SUPERVISOR_TOKEN — the
-            # helpers below also catch the missing-token ``RuntimeError``
-            # from ``make_supervisor_httpx_client`` as defense-in-depth,
-            # so we only still need the sidecar-shape ``server is None``
-            # guard.
-            if server is None:
-                # Addon mode without a live server means we're in the
-                # stdio sidecar — but addon detection should already be
-                # False there. Defensive guard for type-checker + future
-                # refactors.
-                return _bad_request(
-                    "Backup settings POST requires a live MCP server",
-                    code=ErrorCode.INTERNAL_ERROR,
-                    status=500,
-                )
-            # Merge ``clean`` into the *full* current options before posting.
-            # Supervisor validates against the addon schema and rejects any
-            # body missing a required key (notably ``backup_hint`` on the
-            # production / dev manifests). A previous version of this code
-            # shipped just the auto-backup fields, producing
-            # ``addon_configuration_invalid_error: Missing option
-            # 'backup_hint' in root`` from supervisor and a confusing 400 in
-            # the UI even though the user only changed unrelated fields.
-            ok, sup_err = await _supervisor._supervisor_merge_and_post_options(
-                server.settings.verify_ssl, clean
-            )
-            if not ok:
-                assert sup_err is not None
-                logger.warning(
-                    "Supervisor backup-config update failed (%s): %s",
-                    sup_err.kind,
-                    sup_err.message,
-                )
-                code = (
-                    ErrorCode.CONNECTION_FAILED
-                    if sup_err.kind == "transport"
-                    else ErrorCode.CONFIG_VALIDATION_FAILED
-                )
-                return JSONResponse(
-                    create_error_response(code, sup_err.message),
-                    status_code=sup_err.status_code,
-                )
-            # Unified restart flow — see _save_feature_flags for the
-            # rationale. Don't auto-restart from a save handler; the
-            # global Restart Add-on button is the single restart path.
-            return JSONResponse(
-                {
-                    "success": True,
-                    "applied": clean,
-                    "mode": "addon",
-                    "restart_required": True,
-                }
-            )
-
-        # Standalone (file) mode — refuse to override env-pinned fields.
-        rejected: list[dict[str, str]] = []
-        for field_name in list(clean.keys()):
-            env_name = next(
-                en for fn, en, _ in BACKUP_OVERRIDE_FIELDS if fn == field_name
-            )
-            if os.environ.get(env_name) is not None:
-                rejected.append({"field": field_name, "env_var": env_name})
-                del clean[field_name]
-        if rejected:
-            return JSONResponse(
-                {
-                    "success": False,
-                    "error": {
-                        "code": "env_var_pinned",
-                        "message": (
-                            "Some fields are set via environment variable and "
-                            "cannot be changed from the web UI. Unset the "
-                            "env var(s) and reload."
-                        ),
-                        "rejected": rejected,
-                    },
-                },
-                status_code=409,
-            )
-        if not clean:
-            return _bad_request("No editable auto-backup fields after env-var filter")
-        current = _persistence._load_backup_settings_override()
-        current.update(clean)
-        if not _persistence._save_backup_settings_override(current):
-            return JSONResponse(
-                create_error_response(
-                    ErrorCode.CONFIG_VALIDATION_FAILED,
-                    "Failed to persist override file",
-                ),
-                status_code=500,
-            )
-        # Drop the cached Settings so the next read sees the merged value.
-        # File-mode auto-backup settings take effect immediately on the
-        # next ``get_global_settings()`` read — no restart needed, hence
-        # ``restart_required=False``. The JS uses the same field name on
-        # every save endpoint to decide whether to surface the
-        # restart-required banner.
-        _reset_global_settings()
-        return JSONResponse(
-            {
-                "success": True,
-                "applied": clean,
-                "mode": "file",
-                "restart_required": False,
-            }
-        )
-
     handlers: dict[str, Any] = {
         "root_page": _root_page,
         "settings_page": _settings_page,
@@ -2112,14 +1704,6 @@ def build_settings_handlers(
         "save_feature_flags": _save_feature_flags,
         "get_advanced_settings": _get_advanced_settings,
         "save_advanced_settings": _save_advanced_settings,
-        "list_backups": _list_backups,
-        "view_backup": _view_backup,
-        "diff_backup": _diff_backup,
-        "restore_backup": _restore_backup,
-        "delete_backup": _delete_backup,
-        "delete_backups_bulk": _delete_backups_bulk,
-        "get_backup_config": _get_backup_config,
-        "save_backup_config": _save_backup_config,
     }
 
     # Tool security policies. The main server attaches an
@@ -2148,6 +1732,7 @@ def build_settings_handlers(
     handlers.update(build_theme_handlers())
     handlers.update(build_fs_handlers(server))
     handlers.update(build_tools_handlers(server))
+    handlers.update(build_backups_handlers(server))
 
     return handlers
